refactor: combine related plugin files
